requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
rdkit>=2023.3.1
datasets>=2.14.0
huggingface-hub>=0.17.0
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
optuna>=3.3.0
